JS Hooks - Notes & Exercises - Organizing Hooks( put all your JavaScript hooks in a separate file named 'hooks.js'
